+++ test/client_test.js
@@ -39,28 +39,6 @@ describe('Client', function() {
   });
 
   /**
-   * @test {Client#onRequest}
-   */
-  describe('#onRequest', () => {
-    it('should return an `Observable` instead of the underlying `Subject`', () => {
-      let stream = new Client().onRequest;
-      expect(stream).to.be.instanceof(Observable);
-      expect(stream).to.not.be.instanceof(Subject);
-    });
-  });
-
-  /**
-   * @test {Client#onResponse}
-   */
-  describe('#onResponse', () => {
-    it('should return an `Observable` instead of the underlying `Subject`', () => {
-      let stream = new Client().onResponse;
-      expect(stream).to.be.instanceof(Observable);
-      expect(stream).to.not.be.instanceof(Subject);
-    });
-  });
-
-  /**
    * @test {Client#submitHam}
    */
   describe('#submitHam()', () => {